Checking for erros

+++ controllers/favorites.js
@@ -1,12 +1,22 @@
 const user = require('../models/user');
 const recipe = require('../models/recipe');
 const favorite = require('../models/favorite');
+const createError = require('http-errors');
 
 const gettingAllFavorites = (req, res, next) => {
     // #swagger.tags = ['Favorites']
-    favorite.find().then((result) => {
-        res.send(result);
-    });
+    favorite
+        .find()
+        .then((result) => {
+            if (result.length == 0) {
+                throw createError(404, 'There is not favorites');
+            } else {
+                res.send(result);
+            }
+        })
+        .catch((error) => {
+            next(error);
+        });
 };
 
 const gettingFavoritesRecipesFromOneUser = (req, res, next) => {
@@ -14,9 +24,18 @@ const gettingFavoritesRecipesFromOneUser = (req, res, next) => {
     const lookUp = { username: req.params.username };
 
     user.find(lookUp).then((result) => {
-        favorite.find(lookUp).then((result) => {
-            res.send(result);
-        });
+        favorite
+            .find(lookUp)
+            .then((result) => {
+                if (result.length == 0) {
+                    throw createError(404, 'There is no any favorite for this user');
+                } else {
+                    res.send(result);
+                }
+            })
+            .catch((error) => {
+                next(error);
+            });
     });
 };
 
@@ -27,14 +46,23 @@ const updateUserFavorite = (req, res, next) => {
         username: req.body.username
     };
 
-    favorite.find({ username: usernamePara }).then((result) => {
-        const id = result[0]['_id'];
-        const favorites = result[0]['favorites'];
-        updateData['favorites'] = favorites;
-        favorite.findByIdAndUpdate(id, updateData, () => {
-            res.send('Updated');
+    favorite
+        .find({ username: usernamePara })
+        .then((result) => {
+            if (result.length != 0) {
+                const id = result[0]['_id'];
+                const favorites = result[0]['favorites'];
+                updateData['favorites'] = favorites;
+                favorite.findByIdAndUpdate(id, updateData, () => {
+                    res.send('Updated');
+                });
+            } else {
+                throw createError(404, 'There is no any favorite section for this user');
+            }
+        })
+        .catch((error) => {
+            next(error);
         });
-    });
 };
 
 const addFavoriteToUser = (req, res, next) => {
@@ -44,16 +72,25 @@ const addFavoriteToUser = (req, res, next) => {
 
     //We just want to store the IDs of the recipes we like.
 
-    favorite.find({ username: username }).then((result) => {
-        let fav = result[0]['favorites'];
-        fav.push(recipeId);
-        let favoriteId = result[0]['_id'];
-        let newFavoriteObj = result[0];
+    favorite
+        .find({ username: username })
+        .then((result) => {
+            if (result.length != 0) {
+                let fav = result[0]['favorites'];
+                fav.push(recipeId);
+                let favoriteId = result[0]['_id'];
+                let newFavoriteObj = result[0];
 
-        favorite.findByIdAndUpdate(favoriteId, newFavoriteObj, () => {
-            res.send(fav);
+                favorite.findByIdAndUpdate(favoriteId, newFavoriteObj, () => {
+                    res.send(fav);
+                });
+            } else {
+                throw createError(404, 'There is no any favorite section for this user');
+            }
+        })
+        .catch((error) => {
+            next(error);
         });
-    });
 };
 
 const creatingFavoriteForNewUser = (req, res, next) => {
@@ -72,13 +109,21 @@ const creatingFavoriteForNewUser = (req, res, next) => {
 const deletingFavorites = (req, res, next) => {
     // #swagger.tags = ['Favorites']
     const username = req.params.username;
-    favorite.find({ username: username }).then((result) => {
-        let favoriteId = result[0]['_id'];
-        console.log(favoriteId);
-        favorite.findByIdAndDelete(favoriteId, () => {
-            res.send('DELETED!');
+    favorite
+        .find({ username: username })
+        .then((result) => {
+            if (result.length != 0) {
+                let favoriteId = result[0]['_id'];
+                favorite.findByIdAndDelete(favoriteId, () => {
+                    res.send('DELETED!');
+                });
+            } else {
+                throw createError(404, 'There is no any favorite section for this user');
+            }
+        })
+        .catch((error) => {
+            next(error);
         });
-    });
 };
 
 module.exports = {
